Changed  image.rb
